+++ carp_mobile_sensing/CHANGELOG.md
@@ -4,6 +4,7 @@ This should not result in any functional changes, but it requires any Android ap
 [migrate](https://developer.android.com/jetpack/androidx/migrate) if they're using the original support library. 
    * See Flutter [AndroidX compatibility](https://flutter.dev/docs/development/packages-and-plugins/androidx-compatibility)
 * Fixed error in `PedometerProbe`
+
 ## 0.4.0
 * support for data transformers
    * OMH Geolocation